+++ library/src/main/java/uk/co/alt236/easycursor/objectcursor/ObjectFieldAccessor.java
@@ -33,7 +33,7 @@ class ObjectFieldAccessor<T> implements FieldAccessor {
     @Override
     public int getFieldIndexByName(final String name){
         if (mFieldToIndexMap.containsKey(name)) {
-            return mFieldToIndexMap.get(name).intValue();
+            return mFieldToIndexMap.get(name);
         } else {
             return -1;
         }
